random_read_util now can read uint64_t byte buffers from disk

+++ inc/util_random_read.h
@@ -9,6 +9,6 @@
 // it directly works with the provided buffer, reading appropriate data into it from underlying disk using the block_io_functions
 
 // returns 1 on a successfull read, else returns 0
-int random_read_at(void* buffer, uint32_t buffer_size, uint64_t file_offset, const block_io_ops* block_io_functions);
+int random_read_at(void* buffer, uint64_t buffer_size, uint64_t file_offset, const block_io_ops* block_io_functions);
 
 #endif
